working on decompiling
Added Word ">NAME" to look up the name of a word from an execution
token, XT.  Added function "xtToName" to support ">NAME".
Fixed some comments.
Added Word "DUMP"
Added cfa pointer to the word header and updated funtionon to use that
instead of calculating it from the begining of the header and the length
of the name.
Fixed the use of addr_t. Some functions were declaring variables as
addr_t* when they did not need to be.